feat: adding exception handling to GameMode service

+++ src/main/java/com/luizmedeirosn/futs3/services/GameModeService.java
@@ -1,9 +1,12 @@
 package com.luizmedeirosn.futs3.services;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.InvalidDataAccessApiUsageException;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
@@ -19,6 +22,8 @@ import com.luizmedeirosn.futs3.repositories.GameModeRepository;
 import com.luizmedeirosn.futs3.repositories.ParameterRepository;
 import com.luizmedeirosn.futs3.repositories.PositionParameterRepository;
 import com.luizmedeirosn.futs3.repositories.PositionRepository;
+import com.luizmedeirosn.futs3.services.exceptions.DatabaseException;
+import com.luizmedeirosn.futs3.services.exceptions.EntityNotFoundException;
 
 @Service
 public class GameModeService {
@@ -44,7 +49,12 @@ public class GameModeService {
     }
 
     public GameModeMinDTO findById(Long id) {
-        return new GameModeMinDTO(gameModeRepository.findById(id).get());
+        try {
+            return new GameModeMinDTO(gameModeRepository.findById(id).get());
+
+        } catch (NoSuchElementException e) {
+            throw new EntityNotFoundException("GameMode ID not found");
+        }
     }
 
     public List<AllGameModesProjection> findAllFull() {
@@ -53,43 +63,71 @@ public class GameModeService {
     }
 
     public GameModeDTO findFullById(Long id) {
-        return new GameModeDTO( gameModeRepository.findById(id).get(), gameModeRepository.findFullById(id) );
+        try {
+            return new GameModeDTO( gameModeRepository.findById(id).get(), gameModeRepository.findFullById(id) );
+        
+        } catch (NoSuchElementException e) {
+            throw new EntityNotFoundException("GameMode ID not found");
+        }
     }
 
     public GameModeDTO save(PostGameModeDTO postGameModeDTO) {
-        GameMode newGameMode = new GameMode();
-        newGameMode.setFormationName(postGameModeDTO.getFormationName());
-        newGameMode.setDescription(postGameModeDTO.getDescription());
+        try {
+            GameMode newGameMode = new GameMode();
+            newGameMode.setFormationName(postGameModeDTO.getFormationName());
+            newGameMode.setDescription(postGameModeDTO.getDescription());
+            
+            Set<Position> positions = newGameMode.getPositions();
+            postGameModeDTO
+                .getPositionsParameters()
+                .forEach (
+                    positionParameters -> {
+                        Position pos = positionRepository.findById( positionParameters.getPositionId() ).get();
+                        positions.add(pos);
+                        positionParameters
+                            .getParameters()
+                            .forEach (
+                                parameterWeight -> positionParameterRepository.save ( 
+                                    new PositionParameter( pos, parameterRepository.findById(parameterWeight.getParameterId()).get(), parameterWeight.getWeight() )
+                                )
+                            );
+                    }
+                );
+            gameModeRepository.save(newGameMode);
+            GameModeDTO gameModeDTO = findFullById(newGameMode.getId());
+            return gameModeDTO;
+
+        } catch (NoSuchElementException e) {
+            throw new EntityNotFoundException("GameMode request. IDs not found");
+
+        } catch (InvalidDataAccessApiUsageException e) {
+            throw new EntityNotFoundException("GameMode request. The given IDs must not be null");
         
-        Set<Position> positions = newGameMode.getPositions();
-        postGameModeDTO
-            .getPositionsParameters()
-            .forEach (
-                positionParameters -> {
-                    Position pos = positionRepository.findById( positionParameters.getPositionId() ).get();
-                    positions.add(pos);
-                    positionParameters
-                        .getParameters()
-                        .forEach (
-                            parameterWeight -> positionParameterRepository.save ( 
-                                new PositionParameter( pos, parameterRepository.findById(parameterWeight.getParameterId()).get(), parameterWeight.getWeight() )
-                            )
-                        );
-                }
-            );
-        gameModeRepository.save(newGameMode);
-        GameModeDTO gameModeDTO = findFullById(newGameMode.getId());
-        return gameModeDTO;
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("GameMode request. Unique index, check index or primary key violation");
+        }
     }
 
     public GameModeMinDTO update(Long id, UpdateGameModeDTO updateGameModeDTO) {
-        GameMode gameMode = gameModeRepository.getReferenceById(id);
-        gameMode.updateData(updateGameModeDTO);
-        gameMode = gameModeRepository.save(gameMode);
-        return new GameModeMinDTO(gameMode);
+        try {
+            GameMode gameMode = gameModeRepository.getReferenceById(id);
+            gameMode.updateData(updateGameModeDTO);
+            gameMode = gameModeRepository.save(gameMode);
+            return new GameModeMinDTO(gameMode);
+
+        } catch (jakarta.persistence.EntityNotFoundException e) {
+            // jakarta.persistence.EntityNotFoundException: Unable to find com.luizmedeirosn.futs3.entities.GameMode with id 10
+            throw new EntityNotFoundException("GameMode request. IDs not found");
+        
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("GameMode request. Unique index, check index or primary key violation");
+        }
     }
 
     public void deleteById(Long id) {
+        if (!gameModeRepository.existsById(id)) {
+            throw new EntityNotFoundException("GameMode request. ID not found");
+        }
         gameModeRepository.deleteById(id);
     }
